feat(network): add install method for logging and content negotiation
