+++ network/src/commonMain/kotlin/br/com/marvelkmp/network/service/MarvelApiService.kt
@@ -1,34 +1,17 @@
 package br.com.marvelkmp.network.service
 
 import MarvelKmp.network.BuildConfig.PUBLIC_KEY
-import br.com.marvelkmp.logging.util.Log
 import br.com.marvelkmp.network.utils.generateHash
+import br.com.marvelkmp.network.utils.installContentNegotiation
+import br.com.marvelkmp.network.utils.installLogging
 import io.ktor.client.HttpClient
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.json.Json
 
 val marvelApiService = HttpClient {
     val generateHashReturn = generateHash()
 
-    install(Logging) {
-        logger = object : Logger {
-            override fun log(message: String) {
-                Log.v("HTTP Client", null, message)
-            }
-        }
-        level = LogLevel.HEADERS
-    }
-    install(ContentNegotiation) {
-        json(Json {
-            prettyPrint = true
-            ignoreUnknownKeys = true
-        })
-    }
+    installLogging()
+    installContentNegotiation()
     defaultRequest {
         url {
             url("https://gateway.marvel.com/v1/public/")
