+++ network/src/commonMain/kotlin/br/com/marvelkmp/network/service/SuperheroApiService.kt
@@ -1,30 +1,13 @@
 package br.com.marvelkmp.network.service
 
-import br.com.marvelkmp.logging.util.Log
+import br.com.marvelkmp.network.utils.installContentNegotiation
+import br.com.marvelkmp.network.utils.installLogging
 import io.ktor.client.HttpClient
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.json.Json
 
 val superHeroApiService = HttpClient {
-    install(Logging) {
-        logger = object : Logger {
-            override fun log(message: String) {
-                Log.v("HTTP Client", null, message)
-            }
-        }
-        level = LogLevel.HEADERS
-    }
-    install(ContentNegotiation) {
-        json(Json {
-            prettyPrint = true
-            ignoreUnknownKeys = true
-        })
-    }
+    installLogging()
+    installContentNegotiation()
     defaultRequest {
         url {
             url("https://cdn.jsdelivr.net")
