+++ network/src/commonMain/kotlin/br/com/marvelkmp/network/utils/InstallContentNegotiation.kt
@@ -1,0 +1,15 @@
+package br.com.marvelkmp.network.utils
+
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+fun HttpClientConfig<*>.installContentNegotiation() {
+    install(ContentNegotiation) {
+        json(Json {
+            prettyPrint = true
+            ignoreUnknownKeys = true
+        })
+    }
+}
